requirements.txt
google-analytics-data>=0.18.0
pandas>=2.0.0
